+++ src/menus/file_menu.py
@@ -4,24 +4,36 @@ def run_file_menu():
     while True:
 
         print("""
-===============
-  DATA TOOLKIT
-===============\n
+=================
+    FILE TOOLS
+=================\n
         """)
         file_path = input("Masukkan path file: ")
         info = ft.get_file_info(file_path)
 
         if info["exists"]:
             print("✅ File ditemukan!")
-            print("Nama file:", info["name"])
-            print("Ekstensi file:", info["extension"])
-            print("Path absolut:", info["absolute_path"])
+            print(f"Nama File      : {info['name']}")
+            print(f"Ekstensi       : {info['extension']}")
+            print(f"Path Absolut   : {info['absolute_path']}")
+
+         
+            analysis = ft.get_file_analysis(file_path)
+            print("\n--- File Analysis ---")
+            print(f"Jumlah Baris   : {analysis['line_count']}")
+            print(f"Jumlah Kata    : {analysis['word_count']}")
+            print(f"Ukuran File    : {analysis['file_size']} byte")
+
         else:
             print("❌ File tidak ada atau bukan file.")
 
         # pilihan setelah selesai
         while True:
-            print("\n=== Menu File Selesai ===")
+            print("""
+=================
+MENU FILE SELESAI
+=================\n
+        """)
             print("1. Ulangi File Menu")
             print("0. Kembali ke Main Menu")